requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Open-source model dependencies (optional but recommended):
transformers==4.34.1
torch==2.0.0
numpy==1.24.3
scipy==1.11.4

# High-quality local TTS (optional):
# TTS==0.13.1  # Coqui TTS for XTTS
# pyttsx3==2.90  # System TTS fallback